IPMI transport impl around ipmitool
This allows us to support the lanplus interface until we have a pure-Go
impl.  Also allows us to fallback to ipmitool for the lan interface if needed.

Moved 'raw' based types and tests to use the tool transport.

Implementation of the jessup.Control interface remains as-is for now.

Add chassis status handler to IPMI simulator